E-commerce web site
This project, which I developed using ASP.NET Core, MySQL and Node.js, is an e-commerce platform where users can view products, add them to their carts and place orders. On the backend side, a secure and high-performance structure is created with ASP.NET Core, while product and user data is managed with the MySQL database. Real-time communication and API integrations are also provided with Node.js. This project aims to provide a user-friendly interface and scalable infrastructure.